+++ CHANGELOG.md
@@ -56,6 +56,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add decoder management endpoints [(#751)](https://github.com/wazuh/wazuh-indexer-plugins/pull/751)
 - Add draft policy update endpoint [(#758)](https://github.com/wazuh/wazuh-indexer-plugins/pull/758) [(#762)](https://github.com/wazuh/wazuh-indexer-plugins/pull/762)
 - Create integrations with no rules in the SAP [(#765)](https://github.com/wazuh/wazuh-indexer-plugins/pull/765)
+- Create default policies for user content spaces [(#767)](https://github.com/wazuh/wazuh-indexer-plugins/pull/767)
 
 ### Dependencies
 - Upgrade to Gradle 8.14.3 [(#649)](https://github.com/wazuh/wazuh-indexer-plugins/pull/649)